requirements.txt
flask
torch
torchvision
numpy
scipy
scikit-learn
opencv-python-headless
Pillow